+++ modules/postgresql/variables.tf
@@ -243,3 +243,15 @@ variable "add_additional_secret_manager_credentials" {
   type        = bool
   default     = true
 }
+
+variable "instance_edition" {
+  type        = string
+  default     = "ENTERPRISE"
+  description = "Override the default instance edition (`ENTERPRISE` or `ENTERPRISE_PLUS`)."
+}
+
+variable "retain_backups_on_delete" {
+  description = "When this parameter is set to true, Cloud SQL retains backups of the instance even after the instance is deleted."
+  type        = bool
+  default     = true
+}